+++ robotcode/language_server/robotframework/configuration.py
@@ -19,6 +19,7 @@ class RobotConfig(ConfigBase):
     python_path: List[str] = field(default_factory=list)
     env: Dict[str, str] = field(default_factory=dict)
     variables: Dict[str, Any] = field(default_factory=dict)
+    paths: List[str] = field(default_factory=list)
     output_dir: Optional[str] = None
     output_file: Optional[str] = None
     log_file: Optional[str] = None