+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to the "robotcode" extension will be documented in this file.
 
-## [Unreleased]
-- none so far
+## [Unreleased]
+
+- Introduce setting `robotcode.robot.paths` and correspondend launch config property `paths`
+  - Specifies the paths where robot/robotcode should discover test suites. Corresponds to the 'paths' option of robot
+- Introduce setting `robotcode.robot.variableFiles` and correspondend launch config property `variableFiles`
+  - Specifies the variable files for robotframework. Corresponds to the '--variablefile' option of robot.
+
 ##  0.9.2
 
 - none so far